Priority and color added successfully

+++ lib/widgets/task_list.dart
@@ -1,4 +1,4 @@
-// ignore_for_file: use_key_in_widget_constructors
+// ignore_for_file: prefer_const_constructors
 
 import 'package:flutter/material.dart';
 import 'package:provider/provider.dart';
@@ -17,24 +17,81 @@ class TaskList extends StatelessWidget {
               title: Text(
                 task.title,
                 style: TextStyle(
-                  decoration: task.isCompleted
-                      ? TextDecoration.lineThrough
-                      : TextDecoration.none,
-                ),
+                    decoration: task.isCompleted
+                        ? TextDecoration.lineThrough
+                        : TextDecoration.none,
+                    fontWeight: FontWeight.bold,
+                    letterSpacing: 1.5),
               ),
-              trailing: Checkbox(
-                value: task.isCompleted,
-                onChanged: (value) {
-                  taskProvider.toggleTaskCompletion(index);
-                },
+              trailing: Row(
+                mainAxisSize: MainAxisSize.min,
+                children: [
+                  Container(
+                    width: 10.0,
+                    height: 10.0,
+                    decoration: BoxDecoration(
+                      color: _getPriorityColor(task.priority),
+                      shape: BoxShape.circle,
+                    ),
+                  ),
+                  SizedBox(
+                      width: 8.0), // Add spacing between color dot and checkbox
+                  Checkbox(
+                    value: task.isCompleted,
+                    onChanged: (value) {
+                      taskProvider.toggleTaskCompletion(index);
+                    },
+                  ),
+                ],
               ),
               onLongPress: () {
-                taskProvider.deleteTask(index);
+                _showDeleteConfirmationDialog(context, index, taskProvider);
               },
             );
           },
         );
       },
     );
+  }
+
+  void _showDeleteConfirmationDialog(
+      BuildContext context, int index, TaskProvider taskProvider) {
+    showDialog(
+      context: context,
+      builder: (BuildContext context) {
+        return AlertDialog(
+          title: Text('Delete Task'),
+          content: Text('Are you sure you want to delete this task?'),
+          actions: <Widget>[
+            TextButton(
+              child: Text('Cancel'),
+              onPressed: () {
+                Navigator.of(context).pop();
+              },
+            ),
+            TextButton(
+              child: Text('Delete'),
+              onPressed: () {
+                taskProvider.deleteTask(index);
+                Navigator.of(context).pop();
+              },
+            ),
+          ],
+        );
+      },
+    );
+  }
+
+  Color _getPriorityColor(int priority) {
+    switch (priority) {
+      case 1:
+        return Colors.red;
+      case 2:
+        return Colors.orange;
+      case 3:
+        return Colors.green;
+      default:
+        return Colors.black;
+    }
   }
 }
